scripts/youtube/requirements.txt
pytube==12.1.2
ffmpeg-python==0.2.0
requests==2.28.2
gunicorn==20.1.0